Split machine deployment preps from DeployMachinesApp
