+++ internal/command/deploy/machines.go
@@ -3,7 +3,6 @@ package deploy
 import (
 	"context"
 	"fmt"
-	"strconv"
 	"strings"
 	"time"
 
@@ -15,10 +14,8 @@ import (
 	"github.com/superfly/flyctl/client"
 	"github.com/superfly/flyctl/flaps"
 	"github.com/superfly/flyctl/gql"
-	"github.com/superfly/flyctl/helpers"
 	"github.com/superfly/flyctl/internal/appconfig"
 	"github.com/superfly/flyctl/internal/cmdutil"
-	machcmd "github.com/superfly/flyctl/internal/command/machine"
 	"github.com/superfly/flyctl/internal/logger"
 	"github.com/superfly/flyctl/internal/machine"
 	"github.com/superfly/flyctl/iostreams"
@@ -168,252 +165,6 @@ func NewMachineDeployment(ctx context.Context, args MachineDeploymentArgs) (Mach
 		return nil, err
 	}
 	return md, nil
-}
-
-func (md *machineDeployment) resolveProcessGroupChanges() ProcessGroupsDiff {
-	output := ProcessGroupsDiff{
-		groupsToRemove:        map[string]int{},
-		groupsNeedingMachines: map[string]*appconfig.ProcessConfig{},
-	}
-
-	groupHasMachine := map[string]bool{}
-
-	for _, leasableMachine := range md.machineSet.GetMachines() {
-		mach := leasableMachine.Machine()
-		machGroup := mach.ProcessGroup()
-		groupMatch := ""
-		for name := range md.processConfigs {
-			if machGroup == name {
-				groupMatch = machGroup
-				break
-			}
-		}
-		if groupMatch == "" {
-			output.groupsToRemove[machGroup] += 1
-			output.machinesToRemove = append(output.machinesToRemove, leasableMachine)
-		} else {
-			groupHasMachine[groupMatch] = true
-		}
-	}
-
-	for name, val := range md.processConfigs {
-		if hasMach, ok := groupHasMachine[name]; !ok || !hasMach {
-			output.groupsNeedingMachines[name] = val
-		}
-	}
-
-	return output
-}
-
-func (md *machineDeployment) warnAboutProcessGroupChanges(ctx context.Context, diff ProcessGroupsDiff) {
-	var (
-		io                 = iostreams.FromContext(ctx)
-		colorize           = io.ColorScheme()
-		willAddMachines    = len(diff.groupsNeedingMachines) != 0
-		willRemoveMachines = diff.machinesToRemove != nil
-	)
-
-	if !willAddMachines && !willRemoveMachines {
-		return
-	}
-
-	fmt.Fprintln(io.Out, "Process groups have changed. This will:")
-
-	if willRemoveMachines {
-		bullet := colorize.Red("*")
-		for grp, numMach := range diff.groupsToRemove {
-			pluralS := lo.Ternary(numMach == 1, "", "s")
-			fmt.Fprintf(io.Out, " %s destroy %d \"%s\" machine%s\n", bullet, numMach, grp, pluralS)
-		}
-	}
-	if willAddMachines {
-		bullet := colorize.Green("*")
-
-		for name := range diff.groupsNeedingMachines {
-			fmt.Fprintf(io.Out, " %s create 1 \"%s\" machine\n", bullet, name)
-		}
-	}
-	fmt.Fprint(io.Out, "\n")
-}
-
-func (md *machineDeployment) spawnMachineInGroup(ctx context.Context, groupName string) error {
-	if groupName == "" {
-		// If the group is unspecified, it should have been translated to "app" by this point
-		panic("spawnMachineInGroup requires a non-empty group name. this is a bug!")
-	}
-	fmt.Fprintf(md.io.Out, "No machines in group '%s', launching one new machine\n", md.colorize.Bold(groupName))
-	launchInput := md.launchInputForUpdateOrNew(nil, groupName)
-	newMachineRaw, err := md.flapsClient.Launch(ctx, *launchInput)
-	newMachine := machine.NewLeasableMachine(md.flapsClient, md.io, newMachineRaw)
-	if err != nil {
-		return fmt.Errorf("error creating a new machine machine: %w", err)
-	}
-
-	// FIXME: dry this up with release commands and non-empty update
-	fmt.Fprintf(md.io.ErrOut, "  Created release_command machine %s\n", md.colorize.Bold(newMachineRaw.ID))
-	if md.strategy != "immediate" {
-		err := newMachine.WaitForState(ctx, api.MachineStateStarted, md.waitTimeout)
-		if err != nil {
-			return err
-		}
-	}
-	if md.strategy != "immediate" && !md.skipHealthChecks {
-		err := newMachine.WaitForHealthchecksToPass(ctx, md.waitTimeout)
-		// FIXME: combine this wait with the wait for start as one update line (or two per in noninteractive case)
-		if err != nil {
-			return err
-		} else {
-			md.logClearLinesAbove(1)
-			fmt.Fprintf(md.io.ErrOut, "  Machine %s update finished: %s\n",
-				md.colorize.Bold(newMachine.FormattedMachineId()),
-				md.colorize.Green("success"),
-			)
-		}
-	}
-	return nil
-}
-
-func (md *machineDeployment) DeployMachinesApp(ctx context.Context) error {
-	ctx = flaps.NewContext(ctx, md.flapsClient)
-	if md.restartOnly {
-		return md.restartMachinesApp(ctx)
-	}
-	return md.deployMachinesApp(ctx)
-}
-
-func (md *machineDeployment) restartMachinesApp(ctx context.Context) error {
-	err := md.machineSet.AcquireLeases(ctx, md.leaseTimeout)
-	defer func() {
-		err := md.machineSet.ReleaseLeases(ctx)
-		if err != nil {
-			terminal.Warnf("error releasing leases on machines: %v\n", err)
-		}
-	}()
-	if err != nil {
-		return err
-	}
-
-	machineUpdateEntries := lo.Map(md.machineSet.GetMachines(), func(lm machine.LeasableMachine, _ int) *machineUpdateEntry {
-		return &machineUpdateEntry{
-			leasableMachine: lm,
-			launchInput:     md.launchInputForUpdateOrNew(lm.Machine(), ""),
-		}
-	})
-
-	return md.updateExistingMachines(ctx, machineUpdateEntries)
-}
-
-func (md *machineDeployment) deployMachinesApp(ctx context.Context) error {
-	err := md.runReleaseCommand(ctx)
-	if err != nil {
-		return fmt.Errorf("release command failed - aborting deployment. %w", err)
-	}
-
-	if md.machineSet.IsEmpty() {
-		processGroupMachineDiff := ProcessGroupsDiff{
-			groupsToRemove:        map[string]int{},
-			groupsNeedingMachines: md.processConfigs,
-		}
-		md.warnAboutProcessGroupChanges(ctx, processGroupMachineDiff)
-		for name := range md.processConfigs {
-			if err := md.spawnMachineInGroup(ctx, name); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-
-	err = md.machineSet.AcquireLeases(ctx, md.leaseTimeout)
-	defer func() {
-		err := md.machineSet.ReleaseLeases(ctx)
-		if err != nil {
-			terminal.Warnf("error releasing leases on machines: %v\n", err)
-		}
-	}()
-	if err != nil {
-		return err
-	}
-	md.machineSet.StartBackgroundLeaseRefresh(ctx, md.leaseTimeout, md.leaseDelayBetween)
-
-	processGroupMachineDiff := md.resolveProcessGroupChanges()
-	md.warnAboutProcessGroupChanges(ctx, processGroupMachineDiff)
-
-	// Destroy machines that don't fit the current process groups
-	if err := md.machineSet.RemoveMachines(ctx, processGroupMachineDiff.machinesToRemove); err != nil {
-		return err
-	}
-	for _, mach := range processGroupMachineDiff.machinesToRemove {
-		if err := machcmd.Destroy(ctx, md.app, mach.Machine(), true); err != nil {
-			return err
-		}
-	}
-
-	// Create machines for new process groups
-	for name := range processGroupMachineDiff.groupsNeedingMachines {
-		if err := md.spawnMachineInGroup(ctx, name); err != nil {
-			return err
-		}
-	}
-	if len(processGroupMachineDiff.groupsNeedingMachines) > 0 {
-		fmt.Fprintf(md.io.ErrOut, "Finished launching new machines\n")
-	}
-
-	machineUpdateEntries := lo.Map(md.machineSet.GetMachines(), func(lm machine.LeasableMachine, _ int) *machineUpdateEntry {
-		return &machineUpdateEntry{
-			leasableMachine: lm,
-			launchInput:     md.launchInputForUpdateOrNew(lm.Machine(), ""),
-		}
-	})
-
-	return md.updateExistingMachines(ctx, machineUpdateEntries)
-}
-
-type machineUpdateEntry struct {
-	leasableMachine machine.LeasableMachine
-	launchInput     *api.LaunchMachineInput
-}
-
-func (md *machineDeployment) updateExistingMachines(ctx context.Context, updateEntries []*machineUpdateEntry) error {
-	// FIXME: handle deploy strategy: rolling, immediate, canary, bluegreen
-	fmt.Fprintf(md.io.Out, "Updating existing machines in '%s' with %s strategy\n", md.colorize.Bold(md.app.Name), md.strategy)
-	for _, e := range updateEntries {
-		lm := e.leasableMachine
-		launchInput := e.launchInput
-
-		fmt.Fprintf(md.io.ErrOut, "  Updating %s\n", md.colorize.Bold(lm.FormattedMachineId()))
-		err := lm.Update(ctx, *launchInput)
-		if err != nil {
-			if md.strategy != "immediate" {
-				return err
-			} else {
-				fmt.Printf("Continuing after error: %s\n", err)
-			}
-		}
-
-		if md.strategy != "immediate" {
-			err := lm.WaitForState(ctx, api.MachineStateStarted, md.waitTimeout)
-			if err != nil {
-				return err
-			}
-		}
-
-		if md.strategy != "immediate" && !md.skipHealthChecks {
-			err := lm.WaitForHealthchecksToPass(ctx, md.waitTimeout)
-			// FIXME: combine this wait with the wait for start as one update line (or two per in noninteractive case)
-			if err != nil {
-				return err
-			} else {
-				md.logClearLinesAbove(1)
-				fmt.Fprintf(md.io.ErrOut, "  Machine %s update finished: %s\n",
-					md.colorize.Bold(lm.FormattedMachineId()),
-					md.colorize.Green("success"),
-				)
-			}
-		}
-	}
-
-	fmt.Fprintf(md.io.ErrOut, "  Finished deploying\n")
-	return nil
 }
 
 func (md *machineDeployment) setMachinesForDeployment(ctx context.Context) error {
@@ -571,109 +322,6 @@ func (md *machineDeployment) createReleaseInBackend(ctx context.Context) error {
 	md.releaseId = resp.CreateRelease.Release.Id
 	md.releaseVersion = resp.CreateRelease.Release.Version
 	return nil
-}
-
-func (md *machineDeployment) computeMachineConfigMetadata(config *api.MachineConfig) map[string]string {
-	return lo.Assign(
-		md.defaultMachineMetadata(),
-		lo.OmitBy(config.Metadata, func(k, v string) bool {
-			return isFlyAppsPlatformMetadata(k)
-		}))
-}
-
-func (md *machineDeployment) launchInputForRestart(origMachineRaw *api.Machine) *api.LaunchMachineInput {
-	if origMachineRaw == nil {
-		return nil
-	}
-
-	Config := machine.CloneConfig(origMachineRaw.Config)
-	Config.Metadata = md.computeMachineConfigMetadata(Config)
-
-	return &api.LaunchMachineInput{
-		ID:      origMachineRaw.ID,
-		AppID:   md.app.Name,
-		OrgSlug: md.app.Organization.ID,
-		Config:  Config,
-		Region:  origMachineRaw.Region,
-	}
-}
-
-func (md *machineDeployment) setMachineReleaseData(mConfig *api.MachineConfig) {
-	mConfig.Image = md.img
-	mConfig.Metadata = md.computeMachineConfigMetadata(mConfig)
-}
-
-func (md *machineDeployment) resolveUpdatedMachineConfig(origMachineRaw *api.Machine, forReleaseCommand bool) *api.LaunchMachineInput {
-	if md.restartOnly {
-		return md.launchInputForRestart(origMachineRaw)
-	}
-	if forReleaseCommand {
-		return md.launchInputForReleaseCommand(origMachineRaw)
-	}
-	return md.launchInputForUpdateOrNew(origMachineRaw, "")
-}
-
-func (md *machineDeployment) launchInputForUpdateOrNew(origMachineRaw *api.Machine, processGroup string) *api.LaunchMachineInput {
-	// If machine exists we have to respect its ID, Region and Process Group
-	if origMachineRaw == nil {
-		origMachineRaw = &api.Machine{
-			Region: md.appConfig.PrimaryRegion,
-			Config: &api.MachineConfig{},
-		}
-	} else if processGroup == "" {
-		processGroup = origMachineRaw.Config.ProcessGroup()
-	}
-
-	// Ignore the error because by this point we already check the processGroup exists
-	mConfig, _ := md.appConfig.ToMachineConfig(processGroup)
-	mConfig.Guest = helpers.Clone(origMachineRaw.Config.Guest)
-	md.setMachineReleaseData(mConfig)
-
-	// Mounts needs special treatment:
-	//   * Volumes attached to existings machines can't be swapped by other volumes
-	//   * The only possible operation is to update its destination mount path
-	if len(origMachineRaw.Config.Mounts) != 0 {
-		origMount := origMachineRaw.Config.Mounts[0]
-		if len(mConfig.Mounts) == 0 {
-			terminal.Warnf("Machine %s has a volume attached but fly.toml doesn't have a [mounts] section\n", origMachineRaw.ID)
-		} else if len(mConfig.Mounts) != 0 && mConfig.Mounts[0].Path != origMount.Path {
-			terminal.Warnf(
-				"Updating the mount path for volume %s on machine %s from %s to %s due to fly.toml [mounts] destination value\n",
-				origMount.Volume, origMachineRaw.ID, origMount.Path, mConfig.Mounts[0].Path,
-			)
-		}
-		mConfig.Mounts = origMachineRaw.Config.Mounts
-	} else if len(mConfig.Mounts) != 0 {
-		mConfig.Mounts[0].Volume = md.volumes[0].ID
-	}
-
-	return &api.LaunchMachineInput{
-		ID:      origMachineRaw.ID,
-		AppID:   md.app.Name,
-		OrgSlug: md.app.Organization.ID,
-		Region:  origMachineRaw.Region,
-		Config:  mConfig,
-	}
-}
-
-func (md *machineDeployment) defaultMachineMetadata() map[string]string {
-	res := map[string]string{
-		api.MachineConfigMetadataKeyFlyPlatformVersion: api.MachineFlyPlatformVersion2,
-		api.MachineConfigMetadataKeyFlyReleaseId:       md.releaseId,
-		api.MachineConfigMetadataKeyFlyReleaseVersion:  strconv.Itoa(md.releaseVersion),
-		api.MachineConfigMetadataKeyFlyProcessGroup:    api.MachineProcessGroupApp,
-	}
-	if md.app.IsPostgresApp() {
-		res[api.MachineConfigMetadataKeyFlyManagedPostgres] = "true"
-	}
-	return res
-}
-
-func isFlyAppsPlatformMetadata(key string) bool {
-	return key == api.MachineConfigMetadataKeyFlyPlatformVersion ||
-		key == api.MachineConfigMetadataKeyFlyReleaseId ||
-		key == api.MachineConfigMetadataKeyFlyReleaseVersion ||
-		key == api.MachineConfigMetadataKeyFlyManagedPostgres
 }
 
 func (md *machineDeployment) provisionIpsOnFirstDeploy(ctx context.Context) error {
