+++ internal/command/deploy/machines_deploymachinesapp.go
@@ -1,0 +1,238 @@
+package deploy
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/samber/lo"
+	"github.com/superfly/flyctl/api"
+	"github.com/superfly/flyctl/flaps"
+	"github.com/superfly/flyctl/internal/appconfig"
+	machcmd "github.com/superfly/flyctl/internal/command/machine"
+	"github.com/superfly/flyctl/internal/machine"
+	"github.com/superfly/flyctl/iostreams"
+)
+
+func (md *machineDeployment) DeployMachinesApp(ctx context.Context) error {
+	ctx = flaps.NewContext(ctx, md.flapsClient)
+	if md.restartOnly {
+		return md.restartMachinesApp(ctx)
+	}
+	return md.deployMachinesApp(ctx)
+}
+
+func (md *machineDeployment) restartMachinesApp(ctx context.Context) error {
+	err := md.machineSet.AcquireLeases(ctx, md.leaseTimeout)
+	if err != nil {
+		return err
+	}
+	defer md.machineSet.ReleaseLeases(ctx) // skipcq: GO-S2307
+	md.machineSet.StartBackgroundLeaseRefresh(ctx, md.leaseTimeout, md.leaseDelayBetween)
+
+	machineUpdateEntries := lo.Map(md.machineSet.GetMachines(), func(lm machine.LeasableMachine, _ int) *machineUpdateEntry {
+		return &machineUpdateEntry{
+			leasableMachine: lm,
+			launchInput:     md.launchInputForRestart(lm.Machine()),
+		}
+	})
+
+	return md.updateExistingMachines(ctx, machineUpdateEntries)
+}
+
+func (md *machineDeployment) deployMachinesApp(ctx context.Context) error {
+	err := md.runReleaseCommand(ctx)
+	if err != nil {
+		return fmt.Errorf("release command failed - aborting deployment. %w", err)
+	}
+
+	if err := md.machineSet.AcquireLeases(ctx, md.leaseTimeout); err != nil {
+		return err
+	}
+	defer md.machineSet.ReleaseLeases(ctx) // skipcq: GO-S2307
+	md.machineSet.StartBackgroundLeaseRefresh(ctx, md.leaseTimeout, md.leaseDelayBetween)
+
+	processGroupMachineDiff := md.resolveProcessGroupChanges()
+	md.warnAboutProcessGroupChanges(ctx, processGroupMachineDiff)
+
+	if len(processGroupMachineDiff.machinesToRemove) > 0 {
+		// Destroy machines that don't fit the current process groups
+		if err := md.machineSet.RemoveMachines(ctx, processGroupMachineDiff.machinesToRemove); err != nil {
+			return err
+		}
+		for _, mach := range processGroupMachineDiff.machinesToRemove {
+			if err := machcmd.Destroy(ctx, md.app, mach.Machine(), true); err != nil {
+				return err
+			}
+		}
+	}
+
+	// Create machines for new process groups
+	if len(processGroupMachineDiff.groupsNeedingMachines) > 0 {
+		for name := range processGroupMachineDiff.groupsNeedingMachines {
+			if err := md.spawnMachineInGroup(ctx, name); err != nil {
+				return err
+			}
+		}
+		fmt.Fprintf(md.io.ErrOut, "Finished launching new machines\n")
+	}
+
+	machineUpdateEntries := lo.Map(md.machineSet.GetMachines(), func(lm machine.LeasableMachine, _ int) *machineUpdateEntry {
+		return &machineUpdateEntry{
+			leasableMachine: lm,
+			launchInput:     md.launchInputForUpdateOrNew(lm.Machine(), ""),
+		}
+	})
+
+	return md.updateExistingMachines(ctx, machineUpdateEntries)
+}
+
+type machineUpdateEntry struct {
+	leasableMachine machine.LeasableMachine
+	launchInput     *api.LaunchMachineInput
+}
+
+func (md *machineDeployment) updateExistingMachines(ctx context.Context, updateEntries []*machineUpdateEntry) error {
+	// FIXME: handle deploy strategy: rolling, immediate, canary, bluegreen
+	fmt.Fprintf(md.io.Out, "Updating existing machines in '%s' with %s strategy\n", md.colorize.Bold(md.app.Name), md.strategy)
+	for _, e := range updateEntries {
+		lm := e.leasableMachine
+		launchInput := e.launchInput
+
+		fmt.Fprintf(md.io.ErrOut, "  Updating %s\n", md.colorize.Bold(lm.FormattedMachineId()))
+		err := lm.Update(ctx, *launchInput)
+		if err != nil {
+			if md.strategy != "immediate" {
+				return err
+			} else {
+				fmt.Printf("Continuing after error: %s\n", err)
+			}
+		}
+
+		if md.strategy != "immediate" {
+			err := lm.WaitForState(ctx, api.MachineStateStarted, md.waitTimeout)
+			if err != nil {
+				return err
+			}
+		}
+
+		if md.strategy != "immediate" && !md.skipHealthChecks {
+			err := lm.WaitForHealthchecksToPass(ctx, md.waitTimeout)
+			// FIXME: combine this wait with the wait for start as one update line (or two per in noninteractive case)
+			if err != nil {
+				return err
+			} else {
+				md.logClearLinesAbove(1)
+				fmt.Fprintf(md.io.ErrOut, "  Machine %s update finished: %s\n",
+					md.colorize.Bold(lm.FormattedMachineId()),
+					md.colorize.Green("success"),
+				)
+			}
+		}
+	}
+
+	fmt.Fprintf(md.io.ErrOut, "  Finished deploying\n")
+	return nil
+}
+
+func (md *machineDeployment) spawnMachineInGroup(ctx context.Context, groupName string) error {
+	if groupName == "" {
+		// If the group is unspecified, it should have been translated to "app" by this point
+		panic("spawnMachineInGroup requires a non-empty group name. this is a bug!")
+	}
+	fmt.Fprintf(md.io.Out, "No machines in group '%s', launching one new machine\n", md.colorize.Bold(groupName))
+	launchInput := md.launchInputForUpdateOrNew(nil, groupName)
+	newMachineRaw, err := md.flapsClient.Launch(ctx, *launchInput)
+	newMachine := machine.NewLeasableMachine(md.flapsClient, md.io, newMachineRaw)
+	if err != nil {
+		return fmt.Errorf("error creating a new machine machine: %w", err)
+	}
+
+	// FIXME: dry this up with release commands and non-empty update
+	fmt.Fprintf(md.io.ErrOut, "  Created release_command machine %s\n", md.colorize.Bold(newMachineRaw.ID))
+	if md.strategy != "immediate" {
+		err := newMachine.WaitForState(ctx, api.MachineStateStarted, md.waitTimeout)
+		if err != nil {
+			return err
+		}
+	}
+	if md.strategy != "immediate" && !md.skipHealthChecks {
+		err := newMachine.WaitForHealthchecksToPass(ctx, md.waitTimeout)
+		// FIXME: combine this wait with the wait for start as one update line (or two per in noninteractive case)
+		if err != nil {
+			return err
+		} else {
+			md.logClearLinesAbove(1)
+			fmt.Fprintf(md.io.ErrOut, "  Machine %s update finished: %s\n",
+				md.colorize.Bold(newMachine.FormattedMachineId()),
+				md.colorize.Green("success"),
+			)
+		}
+	}
+	return nil
+}
+
+func (md *machineDeployment) resolveProcessGroupChanges() ProcessGroupsDiff {
+	output := ProcessGroupsDiff{
+		groupsToRemove:        map[string]int{},
+		groupsNeedingMachines: map[string]*appconfig.ProcessConfig{},
+	}
+
+	groupHasMachine := map[string]bool{}
+
+	for _, leasableMachine := range md.machineSet.GetMachines() {
+		mach := leasableMachine.Machine()
+		machGroup := mach.ProcessGroup()
+		groupMatch := ""
+		for name := range md.processConfigs {
+			if machGroup == name {
+				groupMatch = machGroup
+				break
+			}
+		}
+		if groupMatch == "" {
+			output.groupsToRemove[machGroup] += 1
+			output.machinesToRemove = append(output.machinesToRemove, leasableMachine)
+		} else {
+			groupHasMachine[groupMatch] = true
+		}
+	}
+
+	for name, val := range md.processConfigs {
+		if ok := groupHasMachine[name]; !ok {
+			output.groupsNeedingMachines[name] = val
+		}
+	}
+
+	return output
+}
+
+func (md *machineDeployment) warnAboutProcessGroupChanges(ctx context.Context, diff ProcessGroupsDiff) {
+	var (
+		io                 = iostreams.FromContext(ctx)
+		colorize           = io.ColorScheme()
+		willAddMachines    = len(diff.groupsNeedingMachines) != 0
+		willRemoveMachines = diff.machinesToRemove != nil
+	)
+
+	if !willAddMachines && !willRemoveMachines {
+		return
+	}
+
+	fmt.Fprintln(io.Out, "Process groups have changed. This will:")
+
+	if willRemoveMachines {
+		bullet := colorize.Red("*")
+		for grp, numMach := range diff.groupsToRemove {
+			pluralS := lo.Ternary(numMach == 1, "", "s")
+			fmt.Fprintf(io.Out, " %s destroy %d \"%s\" machine%s\n", bullet, numMach, grp, pluralS)
+		}
+	}
+	if willAddMachines {
+		bullet := colorize.Green("*")
+
+		for name := range diff.groupsNeedingMachines {
+			fmt.Fprintf(io.Out, " %s create 1 \"%s\" machine\n", bullet, name)
+		}
+	}
+	fmt.Fprint(io.Out, "\n")
+}
